% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{format_tsv}
\alias{format_tsv}
\title{Format and export tsv files}
\usage{
format_tsv(input_folder, output_folder)
}
\arguments{
\item{input_folder}{Path to the folder containing the input text file}

\item{output_folder}{Path to the folder where processed files will be exported}
}
\value{
A list of processed data frames.
}
\description{
this package is designed to format tsv files in a folder, adjust their content and then export the modified data frames to another folder
}
\examples{
dir.create(file.path("input"))# create input folder
dir.create(file.path("output"))# create output folder
download.file("https://raw.githubusercontent.com/noromalala/data_tsv/master/tsv_data.zip", "input/tsv_data.zip", mode = "wb")
unzip("input/tsv_data.zip", exdir = "input")
input_folder <- "input/sarah"
output_folder <- "output"
processed_data <- format_tsv(input_folder, output_folder)
}
