% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TsvFormater-package.R
\docType{package}
\name{TsvFormater-package}
\alias{TsvFormater}
\alias{TsvFormater-package}
\title{TsvFormater: Loading data}
\description{
this package is desingned to format tsv files in a folder, adjust their content and then export the modified data frames to another folder
}
\author{
\strong{Maintainer}: Noromalala Eliette \email{noromalala@utexas.edu}

}
\keyword{internal}
